requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]>=2.5.0
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
pandas>=2.0.0
openpyxl>=3.1.2
reportlab>=4.2.0
ortools>=9.9.3963
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
